styling: split gamesubmenu and fixed imports
