+++ src/components/Game/GamePage/index.css
