+++ src/components/Game/GameSubMenu/index.css
@@ -1,0 +1,63 @@
+.more {
+  position: absolute;
+  top: 56px;
+  right: 42px;
+  color: var(--secondary);
+  font-size: 25px;
+  width: max-content;
+  height: fit-content;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: space-around;
+  border-radius: 10px;
+  margin: auto;
+  transition: 150ms;
+  text-align: center;
+  padding: 20px;
+}
+
+.dots {
+  position: absolute;
+  right: 46px;
+  top: 64px;
+  width: 25px;
+  z-index: 1;
+  cursor: pointer;
+}
+
+.link {
+  font-family: Rubik;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 14px;
+  line-height: 17px;
+  color: #ffffff;
+  margin: 4px;
+}
+
+.linkTitle {
+  font-family: Rubik;
+  font-style: normal;
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 19px;
+  color: var(--primary);
+}
+
+.more .hidden {
+  display: none;
+  cursor: default;
+}
+
+.more.clicked {
+  background: #1a1b1c;
+  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.45);
+}
+
+.more.clicked > .hidden {
+  opacity: 1;
+  display: block;
+
+  cursor: pointer;
+}
