+++ src/components/Game/index.tsx
@@ -1,0 +1,2 @@
+export { default as GamePage } from './GamePage'
+export { default as GameSubMenu } from './GameSubMenu'
